:recycle: refactor: 공통으로 사용되는 클래스를 common 패키지로 이동
- 전체적으로 사용할 수 있는 customview, dialog, ItemDecoration, BaseViewHolder 을 common 패키지로 이동

Issue: #29
