+++ app/src/main/java/com/bibbidi/habittracker/ui/common/dialog/GoalTimePickerBottomSheetDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.bibbidi.habittracker.ui.addhabit.dialog
+package com.bibbidi.habittracker.ui.common.dialog
 
 import android.content.DialogInterface
 import android.os.Bundle
