+++ app/src/main/java/com/bibbidi/habittracker/ui/home/HomeFragment.kt
@@ -16,7 +16,7 @@ import androidx.recyclerview.widget.RecyclerView
 import androidx.viewpager2.widget.ViewPager2
 import com.bibbidi.habittracker.R
 import com.bibbidi.habittracker.databinding.FragmentHomeBinding
-import com.bibbidi.habittracker.ui.ItemDecoration
+import com.bibbidi.habittracker.ui.common.ItemDecoration
 import com.bibbidi.habittracker.ui.model.date.DateItem
 import com.bibbidi.habittracker.ui.model.habit.HabitTypeUiModel
 import com.bibbidi.habittracker.ui.model.habit.log.HabitLogUiModel
