+++ app/src/main/res/layout/shimmer_row_calendar.xml
@@ -11,7 +11,7 @@
         android:layout_height="match_parent"
         android:layout_gravity="center">
 
-        <com.bibbidi.habittracker.ui.customview.DateView
+        <com.bibbidi.habittracker.ui.common.customview.DateView
             android:id="@+id/date_view_sun"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -21,7 +21,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
 
-        <com.bibbidi.habittracker.ui.customview.DateView
+        <com.bibbidi.habittracker.ui.common.customview.DateView
             android:id="@+id/date_view_mon"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -32,7 +32,7 @@
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintWidth_percent="0.14" />
 
-        <com.bibbidi.habittracker.ui.customview.DateView
+        <com.bibbidi.habittracker.ui.common.customview.DateView
             android:id="@+id/date_view_tue"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -42,7 +42,7 @@
             app:layout_constraintStart_toEndOf="@+id/date_view_mon"
             app:layout_constraintTop_toTopOf="parent" />
 
-        <com.bibbidi.habittracker.ui.customview.DateView
+        <com.bibbidi.habittracker.ui.common.customview.DateView
             android:id="@+id/date_view_wed"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -52,7 +52,7 @@
             app:layout_constraintStart_toEndOf="@+id/date_view_tue"
             app:layout_constraintTop_toTopOf="parent" />
 
-        <com.bibbidi.habittracker.ui.customview.DateView
+        <com.bibbidi.habittracker.ui.common.customview.DateView
             android:id="@+id/date_view_thu"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -62,7 +62,7 @@
             app:layout_constraintStart_toEndOf="@+id/date_view_wed"
             app:layout_constraintTop_toTopOf="parent" />
 
-        <com.bibbidi.habittracker.ui.customview.DateView
+        <com.bibbidi.habittracker.ui.common.customview.DateView
             android:id="@+id/date_view_fri"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -72,7 +72,7 @@
             app:layout_constraintStart_toEndOf="@+id/date_view_thu"
             app:layout_constraintTop_toTopOf="parent" />
 
-        <com.bibbidi.habittracker.ui.customview.DateView
+        <com.bibbidi.habittracker.ui.common.customview.DateView
             android:id="@+id/date_view_sat"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
